+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- Unreleased
+    - Optimize repeat matrix
+
 - 2.0.3
     - Matrix decomp methods now return decomp objects
     - Optimized matrix statistical operations